+++ kickFlip-project/src/components/filterOption/filterOption.tsx
@@ -9,7 +9,7 @@ interface FilterOptionProps {
 
 function FilterOption({ attribute, value, handleClick }: FilterOptionProps): JSX.Element {
     return (
-        <li className="accordion-content-item" key={attribute}>
+        <li key={attribute}>
             <button type="button" className="accordion-content-button" onClick={handleClick} data-attribute={attribute}>
                 {attribute === 'color' && (
                     <span className="accordion-content-button-color" style={{ backgroundColor: value }} />
